requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
Django==3.2.7
djangorestframework==3.13.1
hashids==1.3.1
Pillow==8.3.2
psycopg-binary==3.0.11
psycopg2==2.9.1
pymemcache==3.5.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
